Remove step interface and move Build interface
